% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS1000_generate_simulated_samples.R
\name{cnorm}
\alias{cnorm}
\title{Returns the critical value(s) from the normal distribution}
\usage{
cnorm(p, mean = 0, sd = 1, side = c("upper", "both", "lower"))
}
\arguments{
\item{p}{The probability associated with the critical value(s).}

\item{mean}{The mean of the normal distribution. Default is 0.}

\item{sd}{The standard deviation of the normal distribution. Default is 1.}

\item{side}{A string that must be either "upper", "both", or "lower" indicating whether the
'   area associated with the probability p is in the upper tail, both tails, or lower tail.}
}
\description{
This function returns the critical value(s) from the normal distribution
given a central probability. This is particular useful for creating confidence
intervals.
}
\examples{
 # Get the critical value for a 95\% confidence interval from the standard normal distribution
 cnorm(0.95, side = "both")
 
}
