% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS1000_generate_simulated_samples.R
\name{ct}
\alias{ct}
\title{Returns the critical value(s) from a t-distribution}
\usage{
ct(p, df, side = c("upper", "both", "lower"))
}
\arguments{
\item{p}{The probability associated with the critical value(s).}

\item{df}{The degrees of freedom of the t-distribution.}

\item{side}{A string that must be either "upper", "both", or "lower" indicating whether the
'   area associated with the probability p is in the upper tail, both tails, or lower tail.}
}
\description{
This function returns the critical value(s) from a t-distribution
given a central probability and degrees of freedom. This is particular useful for creating confidence
intervals.
}
\examples{
 # Get the critical value for a 95\% confidence interval from the t-distribution with 10 degrees of freedom
 ct(0.95, df = 10, side = "both")
}
