% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{download_github_directory}
\alias{download_github_directory}
\title{Download files from a GitHub directory}
\usage{
download_github_directory(download_dir_name)
}
\arguments{
\item{github_directory_name}{A string indicating the name of the directory in
the class GitHub repository}
}
\description{
This function downloads all files from a specified directory in the class
GitHub site
}
\examples{
 # Download the first homework file
 \dontrun{download_github_directory("homework/homework_01")}

}
