% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{download_practice_session}
\alias{download_practice_session}
\title{Download material from a practice session}
\usage{
download_practice_session(session_number)
}
\arguments{
\item{session_number}{A number indicating which practice session we should get the code
for.}
}
\description{
This function downloads practice session material from the class GitHub
repository.
}
\examples{
 # Download code from the first practice session
 \dontrun{download_practice_session(1)}

}
