% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS1000_generate_simulated_samples.R
\name{get_approval_sample}
\alias{get_approval_sample}
\title{Generate a random sample of a president's approval ratings}
\usage{
get_approval_sample(n, degree_of_approval = FALSE)
}
\arguments{
\item{n}{The sample size.}

\item{degree_of_approval}{A Boolean that if set to TRUE will return a vector
with the levels "strongly disapprove", "disapprove",  "approve", "strongly
approve". If this is set to false, then a vector that only has the levels
"disapprove",  "approve" will be returned.}
}
\description{
`r lifecycle::badge("deprecated")`

This function generates a sample of fictional data of people approve
of the president. This function was renamed rsprinkles() to be more consistent 
with other functions that generate random samples.
}
\examples{
 # Generate a sample from 10 individuals
 get_approval_sample(10)

}
