% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_and_package_parameters.R
\name{get_class_material_root_path}
\alias{get_class_material_root_path}
\title{Get the root path for class materials}
\usage{
get_class_material_root_path()
}
\value{
A string with the path to the root of the class materials directory.
}
\description{
This function gets the root path where the class materials are stored.
If the path has not been set, it will be set to a default location.
}
