% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS1000_generate_simulated_samples.R
\name{get_sprinkle_sample}
\alias{get_sprinkle_sample}
\title{Generate a random sample of sprinkle colors}
\usage{
get_sprinkle_sample(n)
}
\arguments{
\item{n}{The sample size.}
}
\description{
`r lifecycle::badge("deprecated")`

This function generates a fictional sample of data of sprinkles. This
function was renamed rsprinkles() to be more consistent with other functions
that generate random samples.
}
\examples{
 # Generate a sample from 10 sprinkle colors
 get_sprinkle_sample(10)

}
