% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{goto_class}
\alias{goto_class}
\title{Go to the directory for a specific class}
\usage{
goto_class(class_number)
}
\arguments{
\item{class_number}{A number indicating which class we should get the code
for.}
}
\description{
This function takes one to the directory of a specific class. If the 
class files have not been downloaded yet, the user is prompted to download them
from the class GitHub repository
}
\examples{
 # Download code from the first class
 \dontrun{goto_class(1)}

}
