% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{goto_directory}
\alias{goto_directory}
\title{Sets the working directory and has RStudio file pane set to that directory
If the directory does not exist, the user is prompted to download the files.}
\usage{
goto_directory(dir_path, download_message = "")
}
\arguments{
\item{dir_path}{A string indicating the path to the directory relative to
the base save path.}

\item{download_message}{A string that will display what files are being downloaded
(e.g., "homework 1").}
}
\description{
Sets the working directory and has RStudio file pane set to that directory
If the directory does not exist, the user is prompted to download the files.
}
\examples{
 # Go to the first homework directory, downloading files if necessary
 \dontrun{goto_directory("homework/homework_01", "homework 1")}

}
