% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{goto_final_project}
\alias{goto_final_project}
\title{Go to the directory for the final project}
\usage{
goto_final_project()
}
\description{
This function takes one to the directory of the final project. If the
final project files have not been downloaded yet, the user is prompted to
download them from the class GitHub repository
}
\examples{
 # Download code from the final project
 \dontrun{goto_final_project()}

}
