% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{goto_homework}
\alias{goto_homework}
\title{Go to the directory for specific homework}
\usage{
goto_homework(homework_number)
}
\arguments{
\item{homework_number}{A number indicating which homework directory to go to.}
}
\description{
This function takes one to the directory of a specific homework. If the 
homework has not been downloaded yet, the user is prompted to download it from
the class GitHub repository.
}
\examples{
 # Download the first homework file
 \dontrun{goto_homework(1)}

}
