% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{goto_practice_session}
\alias{goto_practice_session}
\title{Go to the directory for a specific practice session}
\usage{
goto_practice_session(session_number)
}
\arguments{
\item{session_number}{A number indicating which practice session should go to
for.}
}
\description{
This function takes one to the directory of a specific practice session. If the 
practice ssion files have not been downloaded yet, the user is prompted to 
download them from the class GitHub repository
}
\examples{
 # Download code from the first practice session
 \dontrun{goto_practice_session(1)}

}
