% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_and_zip.R
\name{list_backups}
\alias{list_backups}
\title{List all available backups}
\usage{
list_backups()
}
\description{
This function lists all available backup zip files in the backup directory.
#' @return A character vector of backup zip file names.
#' @examples
#' \dontrun{list_backups()}
}
