% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_class_files.R
\name{list_github_files}
\alias{list_github_files}
\title{Lists the files that are on the class GitHub site in a particular directory}
\usage{
list_github_files(file_dir_name)
}
\arguments{
\item{file_dir_name}{The name of a directory in the GitHub repository.}
}
\description{
Lists the files that are on the class GitHub site in a particular directory
}
\examples{
 # List files in the data directory on GitHub 
 \dontrun{list_class_files("homework/homework_-1")}

}
