% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS1000_plot_distributions.R
\name{plot_f}
\alias{plot_f}
\title{Plots an F-distribution}
\usage{
plot_f(
  df1,
  df2,
  from = NULL,
  to = NULL,
  n = 1000,
  add = FALSE,
  type = "l",
  xlab = expression(F),
  ylab = "Probability Density",
  xlim = NULL,
  ...
)
}
\arguments{
\item{df1}{degrees of freedom for the numerator (non-negative, but can be non-integer).}

\item{df2}{degrees of freedom for the denominator (non-negative, but can be non-integer).}

\item{from}{the starting x-value for the plot. If NULL, it will be set to 0 or -5 sd from the mean.}

\item{to}{the ending x-value for the plot. If NULL, it will be set to the mean plus 5 sd.}

\item{n}{the number of points to use in the plot. Default is 1000.}

\item{add}{whether to add the curve to an existing plot (TRUE) or create a new plot (FALSE). Default is FALSE.}

\item{type}{the type of plot to create. Default is "l" for a line plot.}

\item{xlab}{the label for the x-axis. Default is expression(chi^2).}

\item{ylab}{the label for the y-axis. Default is "Probability Density".}

\item{xlim}{the limits for the x-axis. If NULL, it will be set to c(from, to).}

\item{...}{additional arguments to be passed to the curve function.}
}
\description{
Plots an F-distribution
}
\examples{
plot_f(5, 10)

set.seed(200)
hist(rf(200, 5, 10), prob = TRUE)
plot_f(5, 10, add = TRUE, col = "blue", lwd = 2)
      
}
