% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS1000_ptail_shuffle_resample_pairs.R
\name{pnull}
\alias{pnull}
\title{Calculates the proportion of points in a vector greater than a specified value}
\usage{
pnull(obs_stat, null_dist, lower.tail = TRUE)
}
\arguments{
\item{obs_stat}{An "observed statistic" value which will be used to assess
how many points in the null_dist are greater than this value.}

\item{null_dist}{A vector of data consistent with a null hypothesis.}

\item{lower.tail}{A Boolean indicating whether one should return the
proportion of points less than or equal to the obs_stat value.}
}
\description{
Calculates the proportion of points in a vector greater than a specified value
}
\examples{
 set.seed(100)
 obs_stat <- 3
 null_dist <- rnorm(1000)
 pnull(obs_stat, null_dist)

}
