% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS1000_ptail_shuffle_resample_pairs.R
\name{ptail}
\alias{ptail}
\title{Calculates the proportion of points in a vector more extreme than a specified value}
\usage{
ptail(obs_value, x, lower.tail = TRUE)
}
\arguments{
\item{obs_value}{A value which will be used to assess
how many points in the vector x are greater (or less) than this value.}

\item{x}{A vector of values the observed_value will be compared to}

\item{lower.tail}{A Boolean indicating whether one should return the
proportion of points less than or equal to the obs_value value.}
}
\description{
Calculates the proportion of points in a vector more extreme than a specified value
}
\examples{
 set.seed(100)
 obs_value <- 3
 null_dist <- rnorm(1000)
 ptail(obs_value, null_dist)

}
