% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS1000_generate_simulated_samples.R
\name{rapprovals}
\alias{rapprovals}
\title{Generate a random sample of a president's approval ratings}
\usage{
rapprovals(n, degree_of_approval = FALSE)
}
\arguments{
\item{n}{The sample size.}

\item{degree_of_approval}{A Boolean that if set to TRUE will return a vector
with the levels "strongly disapprove", "disapprove",  "approve", "strongly
approve". If this is set to false, then a vector that only has the levels
"disapprove",  "approve" will be returned.}
}
\description{
This function generates a sample of fictional data of people approve
of the president.
}
\examples{
 # Generate a sample from 10 individuals
 rapprovals(10)

}
