% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS1000_ptail_shuffle_resample_pairs.R
\name{resample_pairs}
\alias{resample_pairs}
\title{Samples with replacement the same rows from two vectors}
\usage{
resample_pairs(vector1, vector2)
}
\arguments{
\item{vector1}{A vector of values.}

\item{vector2}{A vector of values of the same length as vector1.}
}
\value{
Returns a data frame that has three variables: 
 1. The original sample index number
 2. The values that were randomly selected from vector1
 3. The values that were randomly selected from vector2
}
\description{
Samples with replacement the same rows from two vectors
}
\examples{
 vector1 <- 1:26
 vector2 <- letters
 resample_pairs(vector1, vector2)

}
