% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_and_zip.R
\name{restore_from_backup}
\alias{restore_from_backup}
\title{Restore a backup from a zip file}
\usage{
restore_from_backup(backup_file_name, delete_backup_zip_file = TRUE)
}
\description{
This function restores a backup from a specified zip file in the backup directory.
#' @param backup_file_name The name of the backup zip file to restore.
#' @return NULL
}
\examples{
\dontrun{restore_from_backup("homework/homework_-1_20250817_215216.zip")}
}
