% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS1000_generate_simulated_samples.R
\name{rflip_count}
\alias{rflip_count}
\title{Gives a count/proportion from simulating n coin flips}
\usage{
rflip_count(num_flips = 1, prob = 0.5, report_proportion = FALSE)
}
\arguments{
\item{num_flips}{The number of times to flip the coin.}

\item{prob}{The probability of generated a "heads" on each flip.}

\item{report_proportion}{A Boolean that if set to TRUE will return the
proportion of coin flips that were "heads" otherwise it returns the number
of coin flips that were "heads".}
}
\description{
`r lifecycle::badge("deprecated")`

This function was renamed rflip()
}
\examples{
 set.seed(100)
 rflip_count(10)


}
