% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS1000_generate_simulated_samples.R
\name{rsprinkles}
\alias{rsprinkles}
\title{Generate a random sample of sprinkle colors}
\usage{
rsprinkles(n)
}
\arguments{
\item{n}{The sample size.}
}
\description{
This function generates a fictional sample of data of sprinkles.
}
\examples{
 # Generate a sample from 10 sprinkle colors
 rsprinkles(10)

}
