% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_and_zip.R
\name{save_zip_of_all_class_material}
\alias{save_zip_of_all_class_material}
\title{Save a zip file of all class material}
\usage{
save_zip_of_all_class_material()
}
\description{
This function creates a zip file of the entire class material directory, including all subdirectories and files.
}
\examples{
#' \dontrun{save_zip_of_all_class_material()}
}
