% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS1000_ptail_shuffle_resample_pairs.R
\name{shuffle}
\alias{shuffle}
\title{Shuffles the order of the values in a vector}
\usage{
shuffle(v)
}
\arguments{
\item{v}{A vector of data.}
}
\description{
Shuffles the order of the values in a vector
}
\examples{
 v <- 1:9
 shuffle(v)

}
