% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS1000_get_statistics.R
\name{stats_by_group}
\alias{stats_by_group}
\title{Statistics by group}
\usage{
stats_by_group(data_vector, group_vector, stat = mean)
}
\arguments{
\item{data_vector}{A numeric vector of data.}

\item{group_vector}{A factor vector indicating the group membership for each
observation in data_vector.

@param stat A statistic function to compute separately for each group.}
}
\description{
This function computes a summary statistic by group for a given vector of data.
}
\examples{
# Generate some fictional data
set.seed(100)
data_vector <- rnorm(100)
group_vector <- as.factor(sample(c("A", "B", "C"), size = 100, replace = TRUE))
# Compute the mean of data_vector for each group in group_vector
stats_by_group(data_vector, group_vector, stat = mean)
                                    
}
